properly handle unspecified prefetch meta in parquet options
